requirement.txt.txt
python
django
djangorestframework